+++ CHANGELOG.md
@@ -8,7 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
-- 
+- Add `SetA(this Color, float)` extension method
+- Add a few methods to `GameFile` to specify type of file to write
 
 ### Changed
 
@@ -21,6 +22,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
     - All other classes are in the `LMirman.Utilities` namespace
 - Update some extension methods to have `[Pure]` attribute, notifying if return value is unused (these methods have no consequences if return value unused)
 - The `InBounds` extension method now targets the `ICollection` interface allowing it to be used generically instead of just arrays and lists.
+- ⚠️Breaking: Rename `GameFile` events to no longer begin with `On` term
+- Improved documentation of many public methods (especially `GameFile`)
 
 ## [v2.1.0] - 2024-06-25